uintptr_t cast is nicer.
svn:sync-xref-src-repo-rev: r38865

+++ src/VBox/HostDrivers/Support/win/SUPDrv-win.cpp
@@ -1,4 +1,4 @@
-/* $Id: SUPDrv-win.cpp 13858 2008-11-05 13:45:41Z noreply@oracle.com $ */
+/* $Id: SUPDrv-win.cpp 13867 2008-11-05 14:16:33Z knut.osmundsen@oracle.com $ */
 /** @file
  * VBoxDrv - The VirtualBox Support Driver - Windows NT specifics.
  */
@@ -325,7 +325,7 @@ NTSTATUS _stdcall VBoxDrvNtDeviceControl(PDEVICE_OBJECT pDevObj, PIRP pIrp)
         Assert(KeGetCurrentIrql() <= DISPATCH_LEVEL);
         KIRQL oldIrql;
         KeRaiseIrql(DISPATCH_LEVEL, &oldIrql);
-        int rc = supdrvIOCtlFast(ulCmd, (unsigned)pIrp->UserBuffer /* VMCPU id */, pDevExt, pSession);
+        int rc = supdrvIOCtlFast(ulCmd, (uintptr_t)pIrp->UserBuffer /* VMCPU id */, pDevExt, pSession);
         KeLowerIrql(oldIrql);
 
         /* Complete the I/O request. */
